Rename init.sh to before.sh and post.sh to after.sh Fix dump.sh

+++ .docker/template/TYPO3-v9/before.sh
@@ -19,11 +19,15 @@ if [ ! -f /build/Web/typo3conf/AdditionalConfiguration.php ]; then
   touch /build/Web/FIRST_INSTALL
 fi
 
-e_arrow Populating additional typo3conf files
-cp --recursive --force /configuration/template/${TEMPLATE}/typo3conf/* /build/Web/typo3conf/
+if [ ! -f /build/Web/typo3conf/AdditionalConfiguration.php ]; then
+  e_arrow Populating additional typo3conf files
+  cp --recursive --force /configuration/template/${TEMPLATE}/typo3conf/* /build/Web/typo3conf/
+fi
 
-e_arrow Setting up composer file
-cp --force /configuration/template/${TEMPLATE}/composer.json /build/
+if [ ! -f /build/composer.json ]; then
+  e_arrow Setting up composer file
+  cp --force /configuration/template/${TEMPLATE}/composer.json /build/
+fi
 
 e_arrow Fixing permissions
 chmod -R ug+rwX,o+rX /build
